backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Journal OCR backend API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.51.0",
    "baml-py>=0.210.0",
    "fastapi>=0.115.12",
    "google-genai>=1.14.0",
    "instructor>=1.11.3",
    "lancedb>=0.25.3",
    "openai>=1.70.0",
    "pdf2image>=1.17.0",
    "pillow>=11.2.0",
    "polars>=1.35.2",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "tiktoken>=0.12.0",
    "uvicorn>=0.34.2",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["backend*", "core*", "pipeline*"]

[dependency-groups]
dev = [
    "py-spy>=0.4.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-sugar>=1.1.1",
    "ty>=0.0.1a24",
]

[tool.ty.environment]
# Multiple directories (priority order)
root = ["./src", "./src/core", "./src/pipeline", "./src/backend"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
